+++ app/page.tsx
@@ -1,10 +1,10 @@
 import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
-import NewTweet from "./new-tweet";
-import Tweets from "./tweets";
+import NewTweet from "../components/new-tweet";
+import Tweets from "../components/tweets";
 import { ModeToggle } from "@/components/theme-switcher";
-import AuthButtonClient from "./auth-button-client";
+import AuthButtonClient from "../components/auth-button-client";
 
 export const dynamic = "force-dynamic";
